Refactor repeated code in WireProtocol
- Move Messaging_ProcessPayload to common location.
- Update declaration of Messaging_ProcessPayload in Messaging to allow reuse of header.
- Add declarations to WP App interface header.
- Simplified implementation of WP App Interface in STM and NXP platforms.
